Se actualiza Libreria de Integracion de Factura electronica
git-svn-id: http://192.168.0.251:18080/svn/HGInetMiFacturaElectronica/Tronco@2382 30a10932-8c21-4e4d-906a-15261f94da71

+++ HGInetFeAPI/Service References/ServicioAjusteNomina/Reference.cs
@@ -2872,7 +2872,7 @@ namespace HGInetFeAPI.ServicioAjusteNomina {
         private System.Runtime.Serialization.ExtensionDataObject extensionDataField;
         
         [System.Runtime.Serialization.OptionalFieldAttribute()]
-        private int CantidadField;
+        private decimal CantidadField;
         
         [System.Runtime.Serialization.OptionalFieldAttribute()]
         private string CodigoConceptoField;
@@ -2903,7 +2903,7 @@ namespace HGInetFeAPI.ServicioAjusteNomina {
         }
         
         [System.Runtime.Serialization.DataMemberAttribute()]
-        public int Cantidad {
+        public decimal Cantidad {
             get {
                 return this.CantidadField;
             }
