+++ HGInetFeAPI/Service References/ServicioNomina/Reference.cs
@@ -2808,7 +2808,7 @@ namespace HGInetFeAPI.ServicioNomina {
         private System.Runtime.Serialization.ExtensionDataObject extensionDataField;
         
         [System.Runtime.Serialization.OptionalFieldAttribute()]
-        private int CantidadField;
+        private decimal CantidadField;
         
         [System.Runtime.Serialization.OptionalFieldAttribute()]
         private string CodigoConceptoField;
@@ -2839,7 +2839,7 @@ namespace HGInetFeAPI.ServicioNomina {
         }
         
         [System.Runtime.Serialization.DataMemberAttribute()]
-        public int Cantidad {
+        public decimal Cantidad {
             get {
                 return this.CantidadField;
             }
